Extract types to separate files
- A new type `EndpointInternal` was introduced to represent an object that consists of a map of mutable tracks (`TrackContextImpl`).
- `EndpointInternal` and `TrackContextImpl` are in `internal.ts` because these types depend on one another.
- Other types and interfaces were moved to `types.ts`.
- There is no cyclic import dependency.

# Conflicts:
#	src/webrtc/webRTCEndpoint.ts
